Added Date Form
Been having an issue with the date-fns format() returning the previous day. The problem is timezone related due to parsing an incorrectly formatted date.
The default value returned from an input follows the `yyyy-mm-dd` format, but you want new date to parse a `yyyy/mm/dd` value instead. The solution is a little bit of regex.